aoc/src/test/java/com/yaodanzhang/aoc/year2022/Day5Test.java
package com.yaodanzhang.aoc.year2022;

import static com.yaodanzhang.aoc.input.Inputs.inputOf;
import static org.assertj.core.api.Assertions.assertThat;

import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;

class Day5Test {

  private Day5 puzzle;

  @BeforeEach
  void setUp() {
    puzzle = new Day5(inputOf(this.getClass()));
  }

  @Test
  void should_pass_part_1() {
    assertThat(puzzle.part1Result()).isEqualTo("CMZ");
  }

  @Test
  void should_pass_part_2() {
    assertThat(puzzle.part2Result()).isEqualTo("MCD");
  }

  @Test
  void should_print_out_result() {
    Day5 realPuzzle = new Day5(inputOf(Day5.class));
    System.out.println(realPuzzle.part1Result());
    System.out.println(realPuzzle.part2Result());
  }

}